Handle combatants that have no token.
Modifies the functions that retrieve a token from a combatant to account
for no token to be returned. In the case of the TurnMarker,
MoveToCombatant is changed to clear the marker in the case of an empty
token. The updateCombat hook is modified to use optional chaining,
ensuring that no errors are thrown if there's no token.

This improves compatibility with lancer-initiative, which uses a hidden
combatant at the beginning of the tracker to represent no turn
occurring.